fix(book): throw ResourceNotFoundException when deleting a non-existent book

+++ src/main/java/com/pc/pc/service/BookServiceImpl.java
@@ -82,6 +82,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book", id);
+        }
         bookRepository.deleteById(id);
     }
 }
